# Improve SSSE3 and AVX2 encoding speed