@aklomp 2016-10-13
Wow! This is absolutely astounding! I'm blown away by the fact that you were able to realize that degree of speedup. And the whole thing is so wickedly clever! Using saturating subtraction, interpreting the `0xFF` mask value as signed `-1`... this is really really cool. You win this game. My hat is off to you, sir.

I'm obviously honoured to merge this, but one small remark: could you please fix the indentation to use tabs, and could you please insert a blank line before comment lines? You can force-push the result, it will update the merge request automatically. Thanks!