@mayeut 2016-10-13
I'll do that later in the day (probably in 12 hours or so).
To be fair, I had a version that used one shuffle instruction but with many comparisons (same as before in fact) that was faster than the previous version (this lead to encode being a bit faster that decode). Given the link to Wojciech Muła articles in README.md, I went to his site and saw that he already had a faster version using the saturation substraction trick. I just improved it a little with the mask value as signed `-1` trick.
Also, it uses less constants than the latest version by Wojciech Muła and I guess it will benefit 32bits builds (reduce register pressure) but I did not test that.