@mayeut 2016-10-13
I pushed the small style changes. Tell me if that's what you're expecting. There might be 1 line break that you don't wan't between 2 comments. I did it that way because the first refers to the block of 3 statements that follows.