@aklomp 2016-10-13
Yes, that line break is superfluous in my opinion. Otherwise it's ready to merge. Let me know if you want to push a fix.