@mayeut 2016-10-13
Done. I also modified the commit log: the store instruction is not from that inlined function.